requirements.txt
langchain
streamlit
pypdf2
python-dotenv
faiss-cpu
huggingface_hub

